requirements.txt
# Brain Tumor Detection - Core Dependencies

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0

# Data Augmentation
albumentations>=1.3.0

# Machine Learning
scikit-learn>=1.2.0

# Utilities
tqdm>=4.65.0

# Optional: ONNX Support
onnxruntime>=1.14.0
